Move WebGlContext and OffscreenCanvas to libraries
